requirements.txt
eth-brownie
web3
pytest